+++ src/funcs/deploymentsCancelDeployment.ts
@@ -16,10 +16,6 @@ import {
   CancelDeploymentResponseBody$inboundSchema,
 } from "../models/canceldeploymentresponsebody.js";
 import {
-  CancelDeploymentRequest,
-  CancelDeploymentRequest$outboundSchema,
-} from "../models/canceldeploymentservicesdeploymentsarchitecture.js";
-import {
   ConnectionError,
   InvalidRequestError,
   RequestAbortedError,
@@ -27,6 +23,10 @@ import {
   UnexpectedClientError,
 } from "../models/httpclienterrors.js";
 import { ResponseValidationError } from "../models/responsevalidationerror.js";
+import {
+  CancelDeploymentRequest,
+  CancelDeploymentRequest$outboundSchema,
+} from "../models/routesdestination.js";
 import { SDKValidationError } from "../models/sdkvalidationerror.js";
 import { VercelError } from "../models/vercelerror.js";
 import { APICall, APIPromise } from "../types/async.js";