+++ src/models/countpageviewsop.ts
@@ -244,6 +244,7 @@ export type CountPageviewsData1 = {
   messageId: string;
   eventType: string;
   notificationUrl: string;
+  queueRegion: string;
   sandboxSessionId: string;
   sandboxName: string;
   workflowRunId: string;
@@ -594,6 +595,7 @@ export const CountPageviewsData1$inboundSchema: z.ZodType<
     messageId: types.string(),
     eventType: types.string(),
     notificationUrl: types.string(),
+    queueRegion: types.string(),
     sandboxSessionId: types.string(),
     sandboxName: types.string(),
     workflowRunId: types.string(),